[core] Added sync::Timer
