+++ srtcore/queue.cpp
@@ -489,7 +489,7 @@ CSndQueue::~CSndQueue()
     delete m_pSndUList;
 }
 
-void CSndQueue::init(CChannel *c, srt::sync::SyncEvent *t)
+void CSndQueue::init(CChannel *c, srt::sync::Timer *t)
 {
     m_pChannel                 = c;
     m_pTimer                   = t;
@@ -575,7 +575,7 @@ void *CSndQueue::worker(void *param)
         THREAD_PAUSED();
         if (currtime < next_time)
         {
-            self->m_pTimer->wait_until(next_time);
+            self->m_pTimer->sleep_until(next_time);
 
 #if defined(HAI_DEBUG_SNDQ_HIGHRATE)
             self->m_WorkerStats.lSleepTo++;
