+++ srtcore/sync.cpp
@@ -16,162 +16,7 @@
 #include "srt_compat.h"
 
 
-#ifdef USE_STL_CHRONO
-
-
-////////////////////////////////////////////////////////////////////////////////
-//
-// SyncCond (based on stl chrono C++11)
-//
-////////////////////////////////////////////////////////////////////////////////
-
-srt::sync::SyncCond::SyncCond() {}
-
-srt::sync::SyncCond::~SyncCond() {}
-
-
-bool srt::sync::SyncCond::wait_for(UniqueLock& lk, steady_clock::duration timeout)
-{
-    // Another possible implementation is wait_until(steady_clock::now() + timeout);
-    return m_tick_cond.wait_for(lk, timeout) != cv_status::timeout;
-}
-
-
-void srt::sync::SyncCond::wait(UniqueLock& lk)
-{
-    return m_tick_cond.wait(lk);
-}
-
-
-void srt::sync::SyncCond::notify_one()
-{
-    m_tick_cond.notify_one();
-}
-
-void srt::sync::SyncCond::notify_all()
-{
-    m_tick_cond.notify_all();
-}
-
-
-////////////////////////////////////////////////////////////////////////////////
-//
-// SyncEvent (based on stl chrono C++11)
-//
-////////////////////////////////////////////////////////////////////////////////
-
-srt::sync::SyncEvent::SyncEvent() {}
-
-srt::sync::SyncEvent::~SyncEvent() {}
-
-bool srt::sync::SyncEvent::wait_until(steady_clock::time_point tp)
-{
-    // TODO: Add busy waiting
-
-    // using namespace srt_logging;
-    // LOGC(dlog.Note, log << "SyncEvent::wait_until delta="
-    //    << std::chrono::duration_cast<std::chrono::microseconds>(tp - steady_clock::now()).count() << " us");
-    std::unique_lock<std::mutex> lk(m_tick_lock);
-    m_sched_time = tp;
-    return m_tick_cond.wait_until(lk, tp, [this]() { return m_sched_time <= steady_clock::now(); });
-}
-
-bool srt::sync::SyncEvent::wait_for(const steady_clock::duration& rel_time)
-{
-    std::unique_lock<std::mutex> lk(m_tick_lock);
-    return m_tick_cond.wait_for(lk, rel_time) != cv_status::timeout;
-
-    // wait_until(steady_clock::now() + timeout);
-}
-
-bool srt::sync::SyncEvent::wait_for(UniqueLock& lk, const steady_clock::duration& rel_time)
-{
-    return m_tick_cond.wait_for(lk, rel_time) != cv_status::timeout;
-
-    // wait_until(steady_clock::now() + timeout);
-}
-
-
-void srt::sync::SyncEvent::wait(UniqueLock& lk)
-{
-    return m_tick_cond.wait(lk);
-}
-
-void srt::sync::SyncEvent::wait()
-{
-    std::unique_lock<std::mutex> lk(m_tick_lock);
-    return m_tick_cond.wait(lk);
-}
-
-void srt::sync::SyncEvent::interrupt()
-{
-    {
-        ScopedLock lock(m_tick_lock);
-        m_sched_time = steady_clock::now();
-    }
-
-    notify_one();
-}
-
-void srt::sync::SyncEvent::notify_one()
-{
-    m_tick_cond.notify_one();
-}
-
-void srt::sync::SyncEvent::notify_all()
-{
-    m_tick_cond.notify_all();
-}
-
-////////////////////////////////////////////////////////////////////////////////
-//
-// FormatTime (stl chrono C++11)
-//
-////////////////////////////////////////////////////////////////////////////////
-
-std::string srt::sync::FormatTime(const steady_clock::time_point& timestamp)
-{
-    const uint64_t total_us = count_microseconds(timestamp.time_since_epoch());
-    const uint64_t us = total_us % 1000000;
-    const uint64_t total_sec = total_us / 1000000;
-
-    const uint64_t days = total_sec / (60 * 60 * 24);
-    const uint64_t hours = total_sec / (60 * 60) - days * 24;
-
-    const uint64_t minutes = total_sec / 60 - (days * 24 * 60) - hours * 60;
-    const uint64_t seconds = total_sec - (days * 24 * 60 * 60) - hours * 60 * 60 - minutes * 60;
-
-    ostringstream out;
-    if (days)
-        out << days << "D ";
-    out << setfill('0') << setw(2) << hours << ":"
-        << setfill('0') << setw(2) << minutes << ":"
-        << setfill('0') << setw(2) << seconds << "."
-        << setfill('0') << setw(6) << us << " [STD]";
-    return out.str();
-}
-
-std::string srt::sync::FormatTimeSys(const steady_clock::time_point& timestamp)
-{
-    const time_t                   now_s = ::time(NULL); // get current time in seconds
-    const steady_clock::time_point now_timestamp = steady_clock::now();
-    const int64_t                  delta_us = count_microseconds(timestamp - now_timestamp);
-    const uint64_t now_us = count_microseconds(now_timestamp.time_since_epoch());
-    const int64_t                  delta_s =
-        floor((static_cast<int64_t>(now_us % 1000000) + delta_us) / 1000000.0);
-    const time_t tt = now_s + delta_s;
-    struct tm    tm = SysLocalTime(tt); // in seconds
-    char         tmp_buf[512];
-    strftime(tmp_buf, 512, "%X.", &tm);
-
-    ostringstream out;
-    const uint64_t timestamp_us = count_microseconds(timestamp.time_since_epoch());
-    out << tmp_buf << setfill('0') << setw(6) << (timestamp_us % 1000000) << " [SYS]";
-    return out.str();
-}
-
-
-#else
+#ifndef USE_STL_CHRONO
 
 #if defined(_WIN32)
 #define TIMING_USE_QPC
@@ -190,7 +35,7 @@ namespace srt
 namespace sync
 {
 
- void rdtsc(uint64_t& x)
+void rdtsc(uint64_t& x)
 {
 #ifdef IA32
     uint32_t lval, hval;
@@ -444,45 +289,32 @@ srt::sync::SyncEvent::~SyncEvent()
 }
 
 
-bool srt::sync::SyncEvent::wait_until(TimePoint<steady_clock> tp)
+bool srt::sync::SyncEvent::wait_until(const TimePoint<steady_clock>& tp)
 {
     UniqueLock lck(m_tick_lock);
-    // Use class member such that the method can be interrupted by others
-    m_sched_time = tp;
 
     TimePoint<steady_clock> cur_tp = steady_clock::now();
 
-    while (cur_tp < m_sched_time)
-    {
-#if USE_BUSY_WAITING
-#ifdef IA32
-        __asm__ volatile("pause; rep; nop; nop; nop; nop; nop;");
-#elif IA64
-        __asm__ volatile("nop 0; nop 0; nop 0; nop 0; nop 0;");
-#elif AMD64
-        __asm__ volatile("nop; nop; nop; nop; nop;");
-#endif
-#else
-        const uint64_t wait_us = count_microseconds(m_sched_time - cur_tp);
-        // The while loop ensures that (cur_tp < m_sched_time).
-        // Conversion to microseconds may lose precision, therefore check for 0.
-        if (wait_us == 0)
-            return true;
+    if (cur_tp >= tp)
+        return true;
 
-        timeval now;
-        gettimeofday(&now, 0);
-        const uint64_t time_us = now.tv_sec * uint64_t(1000000) + now.tv_usec + wait_us;
-        timespec       timeout;
-        timeout.tv_sec = time_us / 1000000;
-        timeout.tv_nsec = (time_us % 1000000) * 1000;
+    const uint64_t wait_us = count_microseconds(tp - cur_tp);
+    // Conversion to microseconds may lose precision, therefore check for 0.
+    if (wait_us == 0)
+        return true;
 
-        pthread_cond_timedwait(&m_tick_cond, &lck.m_Mutex.m_mutex, &timeout);
-#endif
+    timeval now;
+    gettimeofday(&now, 0);
+    const uint64_t time_us = now.tv_sec * uint64_t(1000000) + now.tv_usec + wait_us;
+    timespec       timeout;
+    timeout.tv_sec = time_us / 1000000;
+    timeout.tv_nsec = (time_us % 1000000) * 1000;
 
-        cur_tp = steady_clock::now();
-    }
+    pthread_cond_timedwait(&m_tick_cond, &lck.m_Mutex.m_mutex, &timeout);
 
-    return cur_tp >= m_sched_time;
+    cur_tp = steady_clock::now();
+
+    return cur_tp >= tp;
 }
 
 void srt::sync::SyncEvent::notify_one()
@@ -492,13 +324,6 @@ void srt::sync::SyncEvent::notify_one()
 
 void srt::sync::SyncEvent::notify_all()
 {
-    pthread_cond_broadcast(&m_tick_cond);
-}
-
-
-void srt::sync::SyncEvent::interrupt()
-{
-    m_sched_time = steady_clock::now();
     pthread_cond_broadcast(&m_tick_cond);
 }
 
@@ -532,4 +357,70 @@ void srt::sync::SyncEvent::wait(UniqueLock& lk)
 }
 
 
+////////////////////////////////////////////////////////////////////////////////
+//
+// Timer
+//
+////////////////////////////////////////////////////////////////////////////////
+
+srt::sync::Timer::Timer()
+{
+}
+
+
+srt::sync::Timer::~Timer()
+{
+}
+
+
+bool srt::sync::Timer::sleep_until(TimePoint<steady_clock> tp)
+{
+    UniqueLock lck(m_event.mutex());
+    // Use class member such that the method can be interrupted by others
+    m_sched_time = tp;
+
+    TimePoint<steady_clock> cur_tp = steady_clock::now();
+
+    while (cur_tp < m_sched_time)
+    {
+#if USE_BUSY_WAITING
+#ifdef IA32
+        __asm__ volatile("pause; rep; nop; nop; nop; nop; nop;");
+#elif IA64
+        __asm__ volatile("nop 0; nop 0; nop 0; nop 0; nop 0;");
+#elif AMD64
+        __asm__ volatile("nop; nop; nop; nop; nop;");
 #endif
+#else
+        m_event.wait_until(m_sched_time);
+#endif
+
+        cur_tp = steady_clock::now();
+    }
+
+    return cur_tp >= m_sched_time;
+}
+
+
+void srt::sync::Timer::interrupt()
+{
+    UniqueLock lck(m_event.mutex());
+    m_sched_time = steady_clock::now();
+    m_event.notify_all();
+}
+
+
+void srt::sync::Timer::notify_one()
+{
+    m_event.notify_one();
+}
+
+void srt::sync::Timer::notify_all()
+{
+    m_event.notify_all();
+}
+
+#endif
+
+
+srt::sync::SyncEvent s_SyncEvent;
